2024-01-11 refresh + New folder for unlisted visuals
In this commit we add a new feature: Unlisted folder which contains all the visuals that were removed from AppSource in the latest refresh. In Unlisted folder. The versioned pbiviz files remain under "PBIVIZ with versions" subfolder.
In addition, "PBIVIZ with versions" subfolder is now available in Certified and Uncertified folders.