"PBIVIZ with guid" subfolders
Added new "PBIVIZ with guid" subfolders for all the latest visual files using their official GUID as their filename. The visual GUID can help identify the visuals in this repository and match them with the actual visuals you have in your Power BI reports.

The new subfolder will be available in main folders: "All Visuals", "Certified", "Uncertified" and "Unlisted".  Every refresh newer versions of the visuals will replace the older versions in "PBIVIZ with guid". Unlisted visuals will move to Unlisted.

"PBIVIZ" folder will include the same visuals as in "PBIVIZ with guid", but with readable filenames. "PBIVIZ with versions" will remain with all visuals and all the versions that were already stored in this repo.